Save kwargs as a dictionary
I searched the repo and I believe the only append to kwargs happened in
this class. Whereas the only setting happens in
`_ArgumentSpecParser._add_kwargs`.

This is still a backwards incompatible change for people treating
kwargs like a true list, when it is now actually a dictionary as far as
the template is concerned.